+++ Contents/Test_2Q/GameObject/TestGameScene/InfoTextObject.h
