+++ Contents/Test_2Q/GameObject/TestGameScene/SunGeneratorObject.cpp
