+++ Contents/Test_2Q/Scenes/MenuScene.cpp
@@ -3,7 +3,7 @@
 #include "Utils/Singleton.h"
 #include "Platform/Input.h"
 #include "Scene/SceneManager.h"
-#include "GameObject/MenuScene/MenuTextObject.h"
+#include <Test_2Q/GameObject/MenuScene/MenuTextObject.h>
 
 void MenuScene::OnEnterImpl()
 {
