+++ DirectX2D_2.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.13.35825.156
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Engine", "Engine\Engine.vcxproj", "{9F8B7D36-E4AA-4C3F-AADE-81FB4BA00603}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Contents", "Contents\Contents.vcxproj", "{72E55F47-4E9C-41F5-962D-5B96A402C2E6}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Content", "Contents\Contents.vcxproj", "{72E55F47-4E9C-41F5-962D-5B96A402C2E6}"
 	ProjectSection(ProjectDependencies) = postProject
 		{9F8B7D36-E4AA-4C3F-AADE-81FB4BA00603} = {9F8B7D36-E4AA-4C3F-AADE-81FB4BA00603}
 	EndProjectSection
